sua loi sign in up

+++ museverse/src/View/Login-SignUp/SignUp.jsx
@@ -40,7 +40,7 @@ export default function SignUp() {
 
         if (!formData.password.trim()) {
             validationErrors.password = "Password is required"
-        } 
+        }
         // else if (formData.password.length < 6) {
         //     validationErrors.password = "Password should be at least 6 char"
         // }
@@ -49,7 +49,7 @@ export default function SignUp() {
 
         if (Object.keys(validationErrors).length === 0) {
             let item = { username: formData.username, email: formData.email, password: formData.password };
-            let result = await fetch("http://localhost:8000/api/signup", {
+            const response = await fetch("http://localhost:8000/api/signup", {
                 method: 'POST',
                 body: JSON.stringify(item),
                 headers: {
@@ -57,34 +57,41 @@ export default function SignUp() {
                     "Accept": 'application/json'
                 }
             })
-            result = await result.json()
-            console.log("sign up result", result);
+            if (response.ok) {
+                let result = await response.json()
+                console.log("sign up result", result);
 
-            if (result.hasOwnProperty('error')) {
-                const Errors = {};
-                Errors.notmatch = 'Email or username is already existed!';
-                setErrors(Errors);
+                if (result.hasOwnProperty('error')) {
+                    const Errors = {};
+                    Errors.notmatch = 'Email or username is already existed!';
+                    setErrors(Errors);
+                } else {
+                    Swal.fire({
+                        background: "#1F1F22",
+                        color: '#EE5566',
+                        title: "Sign up successfully!",
+                        icon: "success",
+                        confirmButtonText: "Go to HomePage",
+                        confirmButtonColor: '#EE5566',
+                        iconColor: '#EE5566'
+                    }).then((res) => {
+                        if (res.isConfirmed) {
+                            localStorage.setItem('user', JSON.stringify(result));
+                            setLogged(true);
+                            navigate('/');
+                        } else {
+                            localStorage.setItem('user', JSON.stringify(result));
+                            setLogged(true);
+                            navigate('/');
+                        }
+                    });
+                }
             } else {
-                Swal.fire({
-                    background: "#1F1F22",
-                    color: '#EE5566',
-                    title: "Sign up successfully!",
-                    icon: "success",
-                    confirmButtonText: "Go to HomePage",
-                    confirmButtonColor: '#EE5566',
-                    iconColor: '#EE5566'
-                }).then((res) => {
-                    if (res.isConfirmed) {
-                        localStorage.setItem('user', JSON.stringify(result));
-                        setLogged(true);
-                        navigate('/');
-                    }else{
-                        localStorage.setItem('user', JSON.stringify(result));
-                        setLogged(true);
-                        navigate('/');
-                    }
-                });
+                const Errors = {};
+                Errors.notmatch = 'Connection Failed, try again later!';
+                setErrors(Errors);
             }
+
         }
 
     }
